Uploading work done

+++ include/linux_parser.h
@@ -43,7 +43,7 @@ enum CPUStates {
 std::vector<std::string> CpuUtilization();
 long Jiffies();
 long ActiveJiffies();
-long ActiveJiffies(int pid);
+double ActiveJiffies(int pid);
 long IdleJiffies();
 
 // Processes
@@ -51,7 +51,7 @@ std::string Command(int pid);
 std::string Ram(int pid);
 std::string Uid(int pid);
 std::string User(int pid);
-long int UpTime(int pid);
+float UpTime(int pid);
 };  // namespace LinuxParser
 
 #endif
